+++ parser_library/test/lsp/lsp_context_macro_operands_test.cpp
@@ -52,8 +52,11 @@ struct lsp_context_macro_operands : public analyzer_fixture
 TEST_F(lsp_context_macro_operands, non_continued)
 {
     auto res = a.context().lsp_ctx->completion(opencode_loc, { 7, 15 }, 0, completion_trigger_kind::invoked);
-    ASSERT_TRUE(std::holds_alternative<const context::macro_definition*>(res));
-    EXPECT_EQ(std::get<const context::macro_definition*>(res), mac);
+    auto md_sec_p = std::get_if<
+        std::pair<const context::macro_definition*, std::vector<std::pair<const context::symbol*, context::id_index>>>>(
+        &res);
+    ASSERT_TRUE(md_sec_p);
+    EXPECT_EQ(md_sec_p->first, mac);
 }
 
 TEST_F(lsp_context_macro_operands, empty_line_after_macro)
@@ -65,19 +68,29 @@ TEST_F(lsp_context_macro_operands, empty_line_after_macro)
 TEST_F(lsp_context_macro_operands, continued_line_edge)
 {
     auto res = a.context().lsp_ctx->completion(opencode_loc, { 10, 15 }, 0, completion_trigger_kind::invoked);
-    ASSERT_TRUE(std::holds_alternative<const context::macro_definition*>(res));
-    EXPECT_EQ(std::get<const context::macro_definition*>(res), mac);
+    auto md_sec_p = std::get_if<
+        std::pair<const context::macro_definition*, std::vector<std::pair<const context::symbol*, context::id_index>>>>(
+        &res);
+    ASSERT_TRUE(md_sec_p);
+    EXPECT_EQ(md_sec_p->first, mac);
 }
 
 TEST_F(lsp_context_macro_operands, second_continued_line_edge)
 {
     auto res = a.context().lsp_ctx->completion(opencode_loc, { 13, 15 }, 0, completion_trigger_kind::invoked);
-    ASSERT_TRUE(std::holds_alternative<const context::macro_definition*>(res));
-    EXPECT_EQ(std::get<const context::macro_definition*>(res), mac);
+    auto md_sec_p = std::get_if<
+        std::pair<const context::macro_definition*, std::vector<std::pair<const context::symbol*, context::id_index>>>>(
+        &res);
+    ASSERT_TRUE(md_sec_p);
+    EXPECT_EQ(md_sec_p->first, mac);
 }
 
 TEST_F(lsp_context_macro_operands, machine_instruction)
 {
     auto res = a.context().lsp_ctx->completion(opencode_loc, { 14, 15 }, 0, completion_trigger_kind::invoked);
-    EXPECT_TRUE(std::holds_alternative<std::monostate>(res));
+    auto md_sec_p = std::get_if<
+        std::pair<const context::macro_definition*, std::vector<std::pair<const context::symbol*, context::id_index>>>>(
+        &res);
+    ASSERT_TRUE(md_sec_p);
+    EXPECT_FALSE(md_sec_p->first);
 }